feat: change ext to .ts
