+++ package.json
@@ -16,10 +16,10 @@
 		"start": "nest start server",
 		"start:server:dev": "nest start server --watch",
 		"start:server:debug": "nest start server --debug --watch",
-		"start:server:prod": "node dist/apps/tldraw-server.mjs",
+		"start:server:prod": "node dist/apps/tldraw-server.js",
 		"start:worker:dev": "nest start worker --watch",
 		"start:worker:debug": "nest start worker --debug --watch",
-		"start:worker:prod": "node dist/apps/tldraw-worker.mjs",
+		"start:worker:prod": "node dist/apps/tldraw-worker.js",
 		"lint": "eslint \"{src}/**/*.ts\" --fix",
 		"test": "jest",
 		"test:watch": "jest --watch",
