--- src/infra/authorization/index.mts
@@ -1,3 +1,0 @@
-export * from './authorization.module.mjs';
-export * from './authorization.service.mjs';
-
