+++ src/infra/authorization/index.ts
@@ -1,0 +1,3 @@
+export * from './authorization.module.js';
+export * from './authorization.service.js';
+
