--- src/infra/logging/index.mts
@@ -1,3 +1,0 @@
-export * from './logger.mjs';
-export * from './logger.module.mjs';
-
