+++ src/infra/logging/index.ts
@@ -1,0 +1,3 @@
+export * from './logger.js';
+export * from './logger.module.js';
+
