+++ src/infra/logging/logger.module.ts
@@ -1,6 +1,6 @@
 
 import { Module } from '@nestjs/common';
-import { Logger } from './logger.mjs';
+import { Logger } from './logger.js';
 
 @Module({
     providers: [Logger],
