--- src/infra/redis/index.mts
@@ -1,3 +1,0 @@
-export * from './redis.module.mjs';
-export * from './redis.service.mjs';
-
