+++ src/infra/redis/index.ts
@@ -1,0 +1,3 @@
+export * from './redis.module.js';
+export * from './redis.service.js';
+
