--- src/infra/storage/index.mts
@@ -1,3 +1,0 @@
-export * from './storage.module.mjs';
-export * from './storage.service.mjs';
-
