+++ src/infra/storage/index.ts
@@ -1,0 +1,3 @@
+export * from './storage.module.js';
+export * from './storage.service.js';
+
