+++ src/infra/storage/storage.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { LoggerModule } from '../logging/logger.module.mjs';
-import { StorageService } from './storage.service.mjs';
+import { LoggerModule } from '../logging/logger.module.js';
+import { StorageService } from './storage.service.js';
 
 @Module({
     imports: [LoggerModule],
