+++ src/modules/worker/domain/worker.service.ts
@@ -2,8 +2,8 @@
 import { Injectable, OnModuleInit } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { createWorker } from '@y/redis';
-import { RedisService } from '../../../infra/redis/redis.service.mjs';
-import { StorageService } from '../../../infra/storage/storage.service.mjs';
+import { RedisService } from '../../../infra/redis/redis.service.js';
+import { StorageService } from '../../../infra/storage/storage.service.js';
 
 @Injectable()
 export class WorkerService implements OnModuleInit {
